refactor: userService interface 분리
