+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,46 +1,12 @@
 package com.techcourse.service;
 
-import com.techcourse.config.DataSourceConfig;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import nextstep.jdbc.DataAccessException;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
-import org.springframework.transaction.support.DefaultTransactionDefinition;
 
-public class UserService {
+public interface UserService {
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
+    User findById(final long id);
 
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-    }
+    void insert(final User user);
 
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        var dataSource = DataSourceConfig.getInstance();
-        var transactionManager = new DataSourceTransactionManager(dataSource);
-        var transactionStatus = transactionManager.getTransaction(new DefaultTransactionDefinition());
-        try {
-            var user = findById(id);
-            user.changePassword(newPassword);
-            userDao.update(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-            transactionManager.commit(transactionStatus);
-        } catch (Exception e) {
-            e.printStackTrace();
-            transactionManager.rollback(transactionStatus);
-            throw new DataAccessException(e);
-        }
-    }
+    void changePassword(final long id, final String newPassword, final String createBy);
 }
